Moved Drawer to layout directory
