+++ app/src/layouts/CoreLayout/components/Drawer/Drawer.scss
