+++ app/src/layouts/CoreLayout/components/Drawer/DrawerHeader.scss
