+++ app/src/layouts/CoreLayout/components/Drawer/DrawerLogin.scss
@@ -13,6 +13,7 @@
 
     &__link-icon {
         margin-right: 8px;
+        font-size: 18px;
         vertical-align: middle;
     }
 }
