+++ app/src/layouts/CoreLayout/components/Drawer/DrawerLogo.scss
