+++ app/src/layouts/CoreLayout/components/Drawer/DrawerNavigation.scss
