+++ app/src/layouts/CoreLayout/components/Drawer/DrawerNavigationItem.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router";
-import Icon from "../Icon";
+import Icon from "../../../../components/Icon";
 import "./DrawerNavigationItem.scss";
 
 const propTypes = {
