+++ app/src/layouts/CoreLayout/components/Drawer/DrawerProfile.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router";
-import Icon from "../Icon";
+import Icon from "../../../../components/Icon";
 import "./DrawerProfile.scss";
 
 const propTypes = {
@@ -11,9 +11,9 @@ const propTypes = {
 
 const DrawerProfile = ({ avatar, name }) => (
     <div className="drawer-profile">
-        <img src={avatar} className="drawer-profile__avatar" alt="avatar" />
+        <img src={avatar} className="drawer-profile__avatar" alt={`${name}'s avatar`} />
         <div className="drawer-profile__links">
-            <a href="#" className="drawer-profile__name">{name}</a>
+            <Link to={`/users/${name}`} className="drawer-profile__name">{name}</Link>
             <Link to="/logout" className="drawer-profile__logout">
                 <Icon name="exit_to_app" className="drawer-profile__logout-icon" />
                 logout
