+++ app/src/layouts/CoreLayout/components/Drawer/DrawerProfile.scss
